+++ studio/src/app/components/editor/styles/deck/app-deck-transition/app-deck-transition.tsx
@@ -13,97 +13,267 @@ export class AppDeckTransition {
   @Event() private transitionChange: EventEmitter<void>;
 
   @State()
-  private selectedTransition: 'slide' | 'fade' | 'none';
+  private selectedAnimation: 'slide' | 'fade' | 'none';
 
-  private timerInterval: NodeJS.Timeout;
-  private timerCounter: number = 0;
+  @State()
+  private selectedDirection: 'horizontal' | 'vertical' | 'papyrus';
+
+  @State()
+  private device: 'desktop' | 'mobile' = 'desktop';
+
+  private timerIntervalDecks: NodeJS.Timeout;
+  private timerCounterDecks: number = 0;
+
+  private timerIntervalPapyrus: NodeJS.Timeout;
+  private timerCounterPapyrus: number = 0;
 
   async componentWillLoad() {
-    await this.initSelectedTransition();
+    await this.initSelectedAnimation();
+    await this.initSelectedDirection();
   }
 
   async componentDidLoad() {
     await this.animateDecks();
+    await this.animatePapyrus();
   }
 
   disconnectedCallback() {
-    if (this.timerInterval) {
-      clearInterval(this.timerInterval);
+    if (this.timerIntervalDecks) {
+      clearInterval(this.timerIntervalDecks);
+    }
+
+    if (this.timerIntervalPapyrus) {
+      clearInterval(this.timerIntervalPapyrus);
     }
   }
 
-  private initSelectedTransition(): Promise<void> {
-    return new Promise<void>((resolve) => {
-      if (!this.deckElement || !this.deckElement.hasAttribute('transition')) {
-        this.selectedTransition = 'slide';
-        resolve();
-        return;
-      }
+  private async initSelectedAnimation() {
+    if (!this.deckElement || !this.deckElement.hasAttribute('animation')) {
+      this.selectedAnimation = 'slide';
+      return;
+    }
 
-      this.selectedTransition = this.deckElement.getAttribute('transition') as 'slide' | 'fade' | 'none';
+    this.selectedAnimation = this.deckElement.getAttribute('animation') as 'slide' | 'fade' | 'none';
+  }
 
-      resolve();
-    });
+  private async initSelectedDirection() {
+    const attribute: 'direction' | 'direction-mobile' = this.device === 'mobile' ? 'direction-mobile' : 'direction';
+
+    if (!this.deckElement || !this.deckElement.hasAttribute(attribute)) {
+      this.selectedDirection = this.device === 'mobile' ? 'papyrus' : 'horizontal';
+      return;
+    }
+
+    this.selectedDirection = this.deckElement.getAttribute(attribute) as 'horizontal' | 'vertical' | 'papyrus';
   }
 
   private async animateDecks() {
-    this.timerInterval = setInterval(async () => {
-      const elements: NodeListOf<HTMLElement> = this.el.querySelectorAll('deckgo-deck');
+    this.timerIntervalDecks = setInterval(async () => {
+      const elements: NodeListOf<HTMLElement> = this.el.querySelectorAll('div.selected deckgo-deck:not(.papyrus)');
 
       if (elements) {
         for (const element of Array.from(elements)) {
-          if (this.timerCounter % 2 === 0) {
-            await (element as any).slideNext();
+          if (element.classList.contains('showcase-direction')) {
+            await this.animateDecksDirection(element);
           } else {
-            await (element as any).slidePrev();
+            await this.animationDecksAnimation(element);
           }
         }
       }
 
-      this.timerCounter++;
-    }, 2000);
+      this.timerCounterDecks++;
+    }, 1400);
   }
 
-  private applyTransition(transition: 'slide' | 'fade' | 'none'): Promise<void> {
-    return new Promise<void>((resolve) => {
-      if (!this.deckElement || !transition) {
-        resolve();
-        return;
+  private async animatePapyrus() {
+    this.timerIntervalPapyrus = setInterval(async () => {
+      const elements: NodeListOf<HTMLElement> = this.el.querySelectorAll('div.selected div.showcase-papyrus');
+
+      if (elements) {
+        for (const element of Array.from(elements)) {
+          this.timerCounterPapyrus = this.timerCounterPapyrus > 19 ? 0 : this.timerCounterPapyrus + 1;
+
+          element.scroll({top: element.offsetHeight * 4 * (this.timerCounterPapyrus / 20), behavior: 'smooth'});
+        }
       }
 
-      this.deckElement.setAttribute('transition', transition);
+      this.timerCounterPapyrus++;
+    }, 500);
+  }
 
-      this.transitionChange.emit();
+  private async animateDecksDirection(element: HTMLElement) {
+    const end: boolean = await (element as any).isEnd();
 
-      resolve();
-    });
+    if (end) {
+      await (element as any).slideTo(0);
+    } else {
+      await (element as any).slideNext();
+    }
+  }
+
+  private async animationDecksAnimation(element: HTMLElement) {
+    if (this.timerCounterDecks % 2 === 0) {
+      await (element as any).slideNext();
+    } else {
+      await (element as any).slidePrev();
+    }
+  }
+
+  private async applyDirection(direction: 'horizontal' | 'vertical' | 'papyrus') {
+    if (!this.deckElement || !direction) {
+      return;
+    }
+
+    await this.goToFirstSlide();
+
+    this.deckElement.setAttribute(this.device === 'mobile' ? 'direction-mobile' : 'direction', direction);
+
+    this.selectedDirection = direction;
+
+    this.transitionChange.emit();
+  }
+
+  private async goToFirstSlide() {
+    const begin: boolean = await (this.deckElement as HTMLDeckgoDeckElement).isBeginning();
+
+    if (begin) {
+      return;
+    }
+
+    await (this.deckElement as HTMLDeckgoDeckElement).slideTo(0);
+
+    if (!this.deckElement.firstElementChild) {
+      return;
+    }
+
+    this.deckElement.firstElementChild.scroll();
+  }
+
+  private async applyAnimation(animation: 'slide' | 'fade' | 'none') {
+    if (!this.deckElement || !animation) {
+      return;
+    }
+
+    this.deckElement.setAttribute('animation', animation);
+
+    this.selectedAnimation = animation;
+
+    this.transitionChange.emit();
+  }
+
+  private async selectDevice($event: CustomEvent) {
+    if ($event && $event.detail) {
+      this.device = $event.detail.value;
+
+      await this.initSelectedDirection();
+    }
   }
 
   render() {
+    return [this.renderDirection(), this.renderAnimation()];
+  }
+
+  private renderDirection() {
     return (
-      <div class="container ion-margin-top ion-margin-bottom">
-        {this.renderDeckItem('slide', 'var(--ion-color-primary)', 'Slide animation', this.selectedTransition === 'slide')}
+      <app-expansion-panel>
+        <ion-label slot="title">Direction</ion-label>
 
-        {this.renderDeckItem('fade', 'var(--ion-color-secondary)', 'Fade effect', this.selectedTransition === 'fade')}
+        {this.renderDirectionDevice()}
 
-        {this.renderDeckItem('none', 'var(--ion-color-tertiary)', 'Instant transition', this.selectedTransition === 'none')}
+        <div class="container ion-margin-bottom">
+          {this.renderDeckItem('direction', 'horizontal', 'horizontal', this.selectedAnimation, 'Horizontal', this.selectedDirection === 'horizontal', () =>
+            this.applyDirection('horizontal')
+          )}
+
+          {this.renderDeckItem('direction', 'vertical', 'vertical', this.selectedAnimation, 'Vertical', this.selectedDirection === 'vertical', () =>
+            this.applyDirection('vertical')
+          )}
+
+          {this.renderShowcasePapyrus('Papyrus', this.selectedDirection === 'papyrus', () => this.applyDirection('papyrus'))}
+        </div>
+      </app-expansion-panel>
+    );
+  }
+
+  private renderDirectionDevice() {
+    return (
+      <div class="device">
+        <ion-segment mode="ios" value={this.device} onIonChange={($event: CustomEvent) => this.selectDevice($event)}>
+          <ion-segment-button value="desktop" mode="ios">
+            <ion-label>Desktop</ion-label>
+          </ion-segment-button>
+          <ion-segment-button value="mobile" mode="ios">
+            <ion-label>Mobile</ion-label>
+          </ion-segment-button>
+        </ion-segment>
       </div>
     );
   }
 
-  private renderDeckItem(transition: 'slide' | 'fade' | 'none', nextSlideBackground: string, text: string, selected: boolean) {
+  private renderAnimation() {
     return (
-      <div class={`item ${transition} ${selected ? 'selected' : ''}`} custom-tappable onClick={() => this.applyTransition(transition)}>
-        <deckgo-deck embedded={true} keyboard={false} transition={transition}>
-          <deckgo-slide-title>
-            <p slot="title">{text}</p>
-          </deckgo-slide-title>
+      <app-expansion-panel>
+        <ion-label slot="title">Animation</ion-label>
+        <div class="container ion-margin-bottom">
+          {this.renderDeckItem('animation', 'horizontal', 'horizontal', 'slide', 'Swipe', this.selectedAnimation === 'slide', () =>
+            this.applyAnimation('slide')
+          )}
 
-          <deckgo-slide-title style={{'--background': `${nextSlideBackground}`, '--color': 'white'}}>
-            <p slot="title">{text}</p>
-          </deckgo-slide-title>
+          {this.renderDeckItem('animation', 'horizontal', 'horizontal', 'fade', 'Fade', this.selectedAnimation === 'fade', () => this.applyAnimation('fade'))}
+
+          {this.renderDeckItem('animation', 'horizontal', 'horizontal', 'none', 'Instant', this.selectedAnimation === 'none', () =>
+            this.applyAnimation('none')
+          )}
+        </div>
+      </app-expansion-panel>
+    );
+  }
+
+  private renderDeckItem(
+    showcase: 'animation' | 'direction',
+    direction: 'horizontal' | 'vertical' | 'papyrus',
+    directionMobile: 'horizontal' | 'vertical' | 'papyrus',
+    animation: 'slide' | 'fade' | 'none',
+    text: string,
+    selected: boolean,
+    action: (selection) => Promise<void>
+  ) {
+    return (
+      <div class={`item ${selected ? 'selected' : ''} showcase-${showcase}`} custom-tappable onClick={action}>
+        <deckgo-deck
+          embedded={true}
+          keyboard={false}
+          animation={animation}
+          direction={direction}
+          directionMobile={directionMobile}
+          class={`showcase-${showcase}`}>
+          {this.renderItems(text, showcase === 'direction' ? 4 : 2, selected, showcase)}
         </deckgo-deck>
       </div>
     );
+  }
+
+  private renderShowcasePapyrus(text: string, selected: boolean, action: (selection) => Promise<void>) {
+    return (
+      <div class={`item ${selected ? 'selected' : ''} showcase-direction`} custom-tappable onClick={action}>
+        <div class="showcase-papyrus">{this.renderItems(text, 4, selected, 'direction')}</div>
+      </div>
+    );
+  }
+
+  private renderItems(text: string, count: 2 | 4, selected: boolean, showcase: 'direction' | 'animation') {
+    return [...new Array(count)].map((_empty, i: number) => {
+      return (
+        <deckgo-slide-title
+          style={{
+            '--background': `${
+              !selected ? 'transparent' : i % 2 > 0 ? `var(--ion-color-${showcase === 'direction' ? 'primary' : 'tertiary'})` : 'transparent'
+            }`,
+            '--color': !selected ? 'inherit' : `${i % 2 > 0 ? `var(--ion-color-${showcase === 'direction' ? 'primary' : 'tertiary'}-contrast)` : 'inherit'}`,
+          }}>
+          <p slot="title">{text}</p>
+        </deckgo-slide-title>
+      );
+    });
   }
 }