+++ studio/src/app/handlers/editor/events/deck/deck-events.handler.tsx
@@ -741,11 +741,25 @@ export class DeckEventsHandler {
         attributes.style = firebase.firestore.FieldValue.delete();
       }
 
-      if (deck.hasAttribute('transition') && deck.getAttribute('transition') !== 'slide') {
-        attributes.transition = deck.getAttribute('transition') as 'slide' | 'fade' | 'none';
+      if (deck.hasAttribute('animation') && deck.getAttribute('animation') !== 'slide') {
+        attributes.animation = deck.getAttribute('animation') as 'slide' | 'fade' | 'none';
       } else if (updateDeck) {
         // @ts-ignore
-        attributes.transition = firebase.firestore.FieldValue.delete();
+        attributes.animation = firebase.firestore.FieldValue.delete();
+      }
+
+      if (deck.hasAttribute('direction') && deck.getAttribute('direction') !== 'horizontal') {
+        attributes.direction = deck.getAttribute('direction') as 'horizontal' | 'vertical' | 'papyrus';
+      } else if (updateDeck) {
+        // @ts-ignore
+        attributes.direction = firebase.firestore.FieldValue.delete();
+      }
+
+      if (deck.hasAttribute('direction-mobile') && deck.getAttribute('direction-mobile') !== 'papyrus') {
+        attributes.directionMobile = deck.getAttribute('direction-mobile') as 'horizontal' | 'vertical' | 'papyrus';
+      } else if (updateDeck) {
+        // @ts-ignore
+        attributes.directionMobile = firebase.firestore.FieldValue.delete();
       }
 
       resolve(attributes);