+++ studio/src/app/models/data/deck.tsx
@@ -48,7 +48,9 @@ export interface DeckMeta {
 
 export interface DeckAttributes {
   style?: string;
-  transition?: 'slide' | 'fade' | 'none';
+  animation?: 'slide' | 'fade' | 'none';
+  direction?: 'horizontal' | 'vertical' | 'papyrus';
+  directionMobile?: 'horizontal' | 'vertical' | 'papyrus';
 }
 
 export interface DeckClone {