+++ webcomponents/core/src/components.d.ts
@@ -8,9 +8,12 @@ import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { DeckdeckgoAttributeDefinition, DeckdeckgoDeckDefinition, DeckdeckgoSlideDefinition } from "@deckdeckgo/types";
 export namespace Components {
     interface DeckgoDeck {
+        "animation": 'slide' | 'fade' | 'none';
         "blockSlide": (block: boolean) => Promise<void>;
         "cloneBackground": boolean;
         "deleteActiveSlide": () => Promise<void>;
+        "direction": 'horizontal' | 'vertical' | 'papyrus';
+        "directionMobile": 'horizontal' | 'vertical' | 'papyrus';
         "doPrint": () => Promise<void>;
         "embedded": boolean;
         "getActiveIndex": () => Promise<number>;
@@ -20,7 +23,6 @@ export namespace Components {
         "initSlideSize": () => Promise<void>;
         "isBeginning": () => Promise<boolean>;
         "isEnd": () => Promise<boolean>;
-        "isMobile": () => Promise<boolean>;
         "keyboard": boolean;
         "lazyLoadAllContent": () => Promise<void[]>;
         "loadBackground": () => Promise<void>;
@@ -33,7 +35,6 @@ export namespace Components {
         "slideTo": (index: number, speed?: number | undefined, emitEvent?: boolean) => Promise<void>;
         "toggleFullScreen": () => Promise<void>;
         "toggleKeyboardAssist": (state: boolean) => Promise<void>;
-        "transition": 'slide' | 'fade' | 'none';
     }
     interface DeckgoPager {
         "activeIndex": number;
@@ -91,7 +92,10 @@ declare global {
 }
 declare namespace LocalJSX {
     interface DeckgoDeck {
+        "animation"?: 'slide' | 'fade' | 'none';
         "cloneBackground"?: boolean;
+        "direction"?: 'horizontal' | 'vertical' | 'papyrus';
+        "directionMobile"?: 'horizontal' | 'vertical' | 'papyrus';
         "embedded"?: boolean;
         "keyboard"?: boolean;
         "onDeckDidLoad"?: (event: CustomEvent<void>) => void;
@@ -106,7 +110,6 @@ declare namespace LocalJSX {
         "onSlidesDidLoad"?: (event: CustomEvent<any>) => void;
         "reveal"?: boolean;
         "revealOnMobile"?: boolean;
-        "transition"?: 'slide' | 'fade' | 'none';
     }
     interface DeckgoPager {
         "activeIndex"?: number;