+++ docs/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 2.0.0 (2020-09-24)
+
+### Breaking Changes
+
+- route `transition` renamed `animation`
+
+### Features
+
+- introduce `papyrus` and `vertical` direction
+- refresh links in header
+
+### Features
+
+- `isMobile` method removed from core
+
 # 1.9.3 (2020-09-17)
 
 ### Features